refactor: replaced classes for packages
