+++ niffler-e-2-e-tests/src/test/java/niffler/jupiter/di/spend/SpendInjector.java
@@ -1,4 +1,4 @@
-package niffler.jupiter.di;
+package niffler.jupiter.di.spend;
 
 import lombok.SneakyThrows;
 import niffler.database.dao.CategoriesDao;
