+++ niffler-e-2-e-tests/src/test/java/niffler/jupiter/di/spend/WithSpend.java
@@ -1,4 +1,4 @@
-package niffler.jupiter.di;
+package niffler.jupiter.di.spend;
 
 import niffler.database.entity.Currency;
 import org.junit.jupiter.api.extension.ExtendWith;
